backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic_settings==2.1.0
python-keycloak==3.8.4